api/requirements.txt
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
idna==3.3
psycopg2-binary==2.9.3
pydantic==1.9.1
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
